Playing around with vol adjusted regression
I want to be able to trade more securities on Quantopian. However, there's a mismatch between the data available to me on Quantopian and the data available I have for the Bayesian research. My potential solution is to ignore individual securities in the regression, and instead focus on the indicator and volatility. Now, the intercept is the hierarchica part of the model, based on level of volatility